cicd/requirements.txt
pytest
gitpython
